Extract hyperlink hover/click to hyperlinks.rs (refs #68)
Moves handle_hyperlinks() out of main.rs into a dedicated hyperlinks.rs
module. Resolves the hovered cell to a pane, reads back the cell's
OSC 8 hyperlink URL, sets the cursor icon, and opens the URL on
Cmd/Ctrl+click for safe schemes (http/https/mailto).

Mirrors wezterm-gui's per-concern module layout under termwindow/.

No behavior change — pure mechanical refactor.
